requirements.txt
# Core bot & async
aiogram>=3.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0

# DB
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
alembic>=1.11.0

# Cache / FSM storage
redis>=4.6.0

# Media & conversions
Pillow>=11.3.0
pyheif>=0.8.0

# Object storage (Yandex S3-compatible)
boto3>=1.28.0

# Excel / data
pandas>=2.0.0
openpyxl>=3.1.0

# (опционально) файловые операции
aiofiles>=23.2.0
